added more error handlings to use typed error
